livepatch: move non-entitlement related functions to own module
